Simplify!
